Ep 18

+++ app/Http/Controllers/Api/V1/TicketController.php
@@ -30,6 +30,8 @@ class TicketController extends ApiController
     {
         try {
             $user = User::findOrFail($request->input('data.relationships.author.data.id'));
+
+            Gate::authorize('store', null);
         } catch (ModelNotFoundException $exception) {
             return $this->ok('User not found', [
                 'error' => 'The provided user id does not exists'
@@ -86,6 +88,8 @@ class TicketController extends ApiController
         try {
             $ticket = Ticket::findOrFail($ticket_id);
 
+            Gate::authorize('replace', $ticket);
+
             $ticket->update($request->mappedAttributes());
 
             return new TicketResource($ticket);
@@ -101,6 +105,9 @@ class TicketController extends ApiController
     {
         try {
             $ticket = Ticket::findOrFail($ticket_id);
+
+            Gate::authorize('delete', $ticket);
+
             $ticket->delete();
 
             return $this->ok('Ticket successfully deleted');
